# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Boat reservation application API.",
        "title": "Harbour Wave API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/admin/bookings/{bookingId}/cancel": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Admin endpoint to force cancel a booking.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Cancel a booking",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Booking ID (UUID)",
                        "name": "bookingId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Cancellation reason",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_harbour-wave_harbour-wave-backend_internal_model.AdminCancelBookingRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_harbour-wave_harbour-wave-backend_internal_model.AdminCancelBookingResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_harbour-wave_harbour-wave-backend_pkg_httpx.ErrorBody"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/github_com_harbour-wave_harbour-wave-backend_pkg_httpx.ErrorBody"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_harbour-wave_harbour-wave-backend_pkg_httpx.ErrorBody"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/github_com_harbour-wave_harbour-wave-backend_pkg_httpx.ErrorBody"
                        }
                    }
                }
            }
        },
        "/admin/bookings/{bookingId}/price": {
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Admin endpoint to manually override the price of a booking.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Override booking price",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Booking ID (UUID)",
                        "name": "bookingId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Override details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_harbour-wave_harbour-wave-backend_internal_model.AdminPriceOverrideRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_harbour-wave_harbour-wave-backend_internal_model.AdminPriceOverrideResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_harbour-wave_harbour-wave-backend_pkg_httpx.ErrorBody"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/github_com_harbour-wave_harbour-wave-backend_pkg_httpx.ErrorBody"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_harbour-wave_harbour-wave-backend_pkg_httpx.ErrorBody"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/github_com_harbour-wave_harbour-wave-backend_pkg_httpx.ErrorBody"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/github_com_harbour-wave_harbour-wave-backend_pkg_httpx.ErrorBody"
                        }
                    }
                }
            }
        },
        "/admin/dates/{date}/block": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Admin endpoint to block bookings for an entire date.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Block an entire date",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Date in YYYY-MM-DD format",
                        "name": "date",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Block reason",
                        "name": "request",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/github_com_harbour-wave_harbour-wave-backend_internal_model.AdminBlockDateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_harbour-wave_harbour-wave-backend_internal_model.AdminBlockDateResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_harbour-wave_harbour-wave-backend_pkg_httpx.ErrorBody"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/github_com_harbour-wave_harbour-wave-backend_pkg_httpx.ErrorBody"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/github_com_harbour-wave_harbour-wave-backend_pkg_httpx.ErrorBody"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/github_com_harbour-wave_harbour-wave-backend_pkg_httpx.ErrorBody"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Admin endpoint to remove a block on an entire date.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Unblock an entire date",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Date in YYYY-MM-DD format",
                        "name": "date",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_harbour-wave_harbour-wave-backend_internal_model.AdminUnblockDateResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_harbour-wave_harbour-wave-backend_pkg_httpx.ErrorBody"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/github_com_harbour-wave_harbour-wave-backend_pkg_httpx.ErrorBody"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_harbour-wave_harbour-wave-backend_pkg_httpx.ErrorBody"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/github_com_harbour-wave_harbour-wave-backend_pkg_httpx.ErrorBody"
                        }
                    }
                }
            }
        },
        "/admin/slots/{date}/{time}/block": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Admin endpoint to block bookings for a specific date and time.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Block a specific slot",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Date in YYYY-MM-DD format",
                        "name": "date",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Time in HH:MM format",
                        "name": "time",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Block reason",
                        "name": "request",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/github_com_harbour-wave_harbour-wave-backend_internal_model.AdminBlockSlotRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_harbour-wave_harbour-wave-backend_internal_model.AdminBlockSlotResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_harbour-wave_harbour-wave-backend_pkg_httpx.ErrorBody"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/github_com_harbour-wave_harbour-wave-backend_pkg_httpx.ErrorBody"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_harbour-wave_harbour-wave-backend_pkg_httpx.ErrorBody"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/github_com_harbour-wave_harbour-wave-backend_pkg_httpx.ErrorBody"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/github_com_harbour-wave_harbour-wave-backend_pkg_httpx.ErrorBody"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Admin endpoint to remove a block on a specific date and time.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Unblock a specific slot",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Date in YYYY-MM-DD format",
                        "name": "date",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Time in HH:MM format",
                        "name": "time",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_harbour-wave_harbour-wave-backend_internal_model.AdminUnblockSlotResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_harbour-wave_harbour-wave-backend_pkg_httpx.ErrorBody"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/github_com_harbour-wave_harbour-wave-backend_pkg_httpx.ErrorBody"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_harbour-wave_harbour-wave-backend_pkg_httpx.ErrorBody"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/github_com_harbour-wave_harbour-wave-backend_pkg_httpx.ErrorBody"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "github_com_harbour-wave_harbour-wave-backend_internal_model.AdminBlockDateRequest": {
            "type": "object",
            "properties": {
                "reason": {
                    "type": "string"
                }
            }
        },
        "github_com_harbour-wave_harbour-wave-backend_internal_model.AdminBlockDateResponse": {
            "type": "object",
            "properties": {
                "blocked": {
                    "type": "boolean"
                },
                "blockedAt": {
                    "type": "string"
                },
                "date": {
                    "type": "string"
                },
                "reason": {
                    "type": "string"
                }
            }
        },
...
        "github_com_harbour-wave_harbour-wave-backend_internal_model.AdminCancelBookingRequest": {
            "type": "object",
            "required": [
                "reason"
            ],
            "properties": {
                "reason": {
                    "type": "string",
                    "minLength": 1
                }
            }
        },
        "github_com_harbour-wave_harbour-wave-backend_internal_model.AdminCancelBookingResponse": {
            "type": "object",
            "properties": {
                "bookingId": {
                    "type": "string"
                },
                "cancelledAt": {
                    "type": "string"
                },
                "reason": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "github_com_harbour-wave_harbour-wave-backend_internal_model.AdminPriceOverrideRequest": {
            "type": "object",
            "required": [
                "reason"
            ],
            "properties": {
                "reason": {
                    "type": "string",
                    "minLength": 5
                },
                "totalAmount": {
                    "description": "nil clears the override",
                    "type": "number"
                }
            }
        },
        "github_com_harbour-wave_harbour-wave-backend_internal_model.AdminPriceOverrideResponse": {
            "type": "object",
            "properties": {
                "bookingId": {
                    "type": "string"
                },
                "effectiveAmount": {
                    "type": "number"
                },
                "overrideReason": {
                    "type": "string"
                },
                "priceOverride": {
                    "type": "number"
                },
                "totalAmount": {
                    "type": "number"
                }
            }
        },
...
        "github_com_harbour-wave_harbour-wave-backend_pkg_httpx.ErrorBody": {
            "type": "object",
            "properties": {
                "detail": {
                    "type": "string"
                },
                "issues": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_harbour-wave_harbour-wave-backend_pkg_httpx.Issue"
                    }
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}